Update README.md
The JS library listed requires adding stylesheets to your app that contain all images as data uris. The only problem with that is that you have to serve 5mb of emoji images to all users of your app. I wrote and added an alternate JS library that just parses text for strings that match and replace them with image tags that point to the actual image. This means you send the browser 14kb of js instead. Much slimmer and the code is quite a bit more efficient as well.